refactor: move bases from package
