+++ grails-app/domain/com/mini/asaas/Customer.groovy
@@ -1,6 +1,6 @@
 package com.mini.asaas
 
-import com.mini.asaas.utils.PersonEntity
+import com.mini.asaas.base.PersonEntity
 import grails.compiler.GrailsCompileStatic
 
 @GrailsCompileStatic
