+++ grails-app/domain/com/mini/asaas/Payer.groovy
@@ -1,6 +1,6 @@
 package com.mini.asaas
 
-import com.mini.asaas.utils.PersonEntity
+import com.mini.asaas.base.PersonEntity
 import com.mini.asaas.Customer
 import grails.compiler.GrailsCompileStatic
