+++ grails-app/domain/com/mini/asaas/payment/Payment.groovy
@@ -4,7 +4,7 @@ import com.mini.asaas.Payer
 import com.mini.asaas.Customer
 import com.mini.asaas.enums.payment.BillingType
 import com.mini.asaas.enums.payment.PaymentStatus
-import com.mini.asaas.utils.BaseEntity
+import com.mini.asaas.base.BaseEntity
 import grails.compiler.GrailsCompileStatic
 
 @GrailsCompileStatic
