+++ src/main/groovy/com/mini/asaas/base/PersonEntity.groovy
@@ -1,8 +1,8 @@
-package com.mini.asaas.utils
+package com.mini.asaas.base
 
 import com.mini.asaas.Address
 import com.mini.asaas.enums.PersonType
-import com.mini.asaas.utils.BaseEntity
+import com.mini.asaas.base.BaseEntity
 import grails.compiler.GrailsCompileStatic
 
 @GrailsCompileStatic
